requirements.txt
asgiref==3.3.1
collection==0.1.6
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sqlparse==0.4.1
stop-words==2018.7.23
whitenoise==5.2.0
wordcloud==1.8.1
